refactor: update analytics integration and remove SEO configuration
- Replaced the existing analytics script generation with direct script tags in astro.config.mjs for improved clarity.
- Removed the seo.js file, consolidating SEO-related functionality into the main configuration file.
